Always wait for block devices to appear
This fixes an issue where the uboot environment load wouldn't wait. The
solution is that everyone waits and this funnels all accesses through
one function to make this happen.

The wait is not cumulative, so if there are a dozen uboot environment
loads, scripts don't get 12*the wait for bad specs. They just get one
and the rest immediately time out.