Refactor commandline argument passing
The Linux kernel intercepts a LOT of different ways of passing
commandline parameters. This is very confusing, but makes sense from the
Linux kernel's perspective. To make things less confusing for
`nerves_initramfs` users, add instructions for passing all commandline
args after a `--` on the commandline. Linux won't process anything after
that. This also allows simplifications to the parameters to just
key=value or just key. See the README.md for examples.