Force a reboot after reverting firmware
This is an opinionated change to avoid a hard-to-debug situation where
the script does not reboot after reverting. In this case, the Linux
kernel could be different. This might cause kernel modules not to load
(due to version mismatches, API changes, etc.), and will certainly be
confusing to debug if it ever happened in the field.